Fixing some ambiguity.
Right now the no_bias_initialization refers to two things: Whether to
push the biases_init configuration, and also whether to accept it or
not. I think the variable should refer to only one of these two cases
(and if both cases need to be supported, they should probably be
represented by different attributes). This way I think it is much
clearer what is going on.