Streamline parameter passing to algorithms
DifferentiableCostMinimizer needs to know which parameters are being
optimized and currently it gets all shared variables from the graph.
This is bad because assumes that all shared variables are trainable
that is not true for constants such as ones in MRG_RandomStream.

Having user always specify parameters improves transparency and
keeps framework modular. For convenience, add new property to
ComputationGraph. To avoid circular dependencies, implement role lookup
in roles.py and make both VariableFilter and ComputationGraph use it.

Fixes #364