redundent sort is now removed
warning if variable is not in computation graph

warning if variable is an auxiliary variable

print name of variable in warning msg.

check AUXILIARY variable test on variable, 'if' to 'elif'

first loop removed and all checks are done in one loop

typo fixed

print list of all remaining_replacements

warning in a cleaner way.

comment on pop() added

unit tests added