ConvolutionalSequence: use num_output_channels.
Define a property num_output_channels, which is really the more
appropriate thing to be querying within ConvolutionalSequence to define
the geometry of the input image.

This change, along with the previous commit, make MaxPooling a
first-class citizen within a ConvolutionalSequence. (The num_channels
and image_size properties are kind of ugly for now but they get the job
done, pending any revisiting of the constructor arguments.)