Rename W, b -> scale, shift.
Cesar makes a good point that using the VariableFilter to select all
weights in a model, you might not want to select batch norm scalings,
etc. I've given them functional names of "scale" and "shift" instead
of obscure Greek letters that you have to have read the paper (and
remember which is which).

BATCH_NORM_SHIFT_PARAMETER inherits from BIAS, since there is no real
difference between it and a normal bias, and you might want to select
all the biases in a BatchNormalizedMLP, and those are the only ones
really. BATCH_NORM_SCALE_PARAMETER simply inherits from PARAMETER,
because it's a bit different in usage and semantics than a normal
linear transformation and plays a different role in neural net models.