Saner use_bias push behaviour.
Push use_bias to children only if a value is explicitly specified
in the parent.  This mirrors the behaviour of weights_init and
biases_init.

This can be accomplished by either Initializable using a value of
`None` (and `Linear` and `ConvolutionalSequence`) special-casing
that value to mean True) or simply by not setting the attribute.
I've chosen the latter since it's more likely to raise an error
if it's not what you expect: `bool(None)` is `False`, and so this might
lead to subtle bugs and unintended behaviour. The relevant classes
instead use `getattr(self, 'use_bias', class_preferred_default)`.