ConvolutionalTranspose: infer original_image_size.
Transposed (strided) convolution is a one-to-many mapping, with zero
padding of border pixels dropped by the original convolution being
reconstructed.

In the most common case, you want to produce an upsampled output image
of which the transposed convolutional filter touches every pixel. This
adds the automatic inference of the original_image_size with this case
in mind. It also adds documentation on what constitutes a valid
original_image_size for manual specification.