Fix some None-handling in ConvolutionalSequence
ConvolutionalSequence breaks when using it with layers that don't have
the image size specified. The reason is due to the constructor setting
the default for image_size to None when it should be (None, None), as is
done in Convolutional's constructor. There is also a None-check that
checks whether hte image size is None rather than a tuple of Nones.