Remove lame duck DifferentiableCostMinimizer.
The all_updates business masks the fact that you really are
modifying self.updates in-place, so I removed it to make it clearer.

The change in graph.py is so that you can pass other iterables into
ComputationGraph (and also so any list object is copied, so that you
can't accidentally modify the internal state of the ComputationGraph).